fix: 7 Move compilation fixes for DLMM library output
Code generation:
- Collapse chained casts: (x as u256) as u128 → (x as u128)
- Strip leading zeros from numeric literals (0.1e18 → 100000000000000000)
- Skip no-op casts where literal suffix already matches target type

Shift operations (Move requires u8 right operand):
- Binary shift: (x << y) → (x << (y as u8))
- Compound shift assignments: (x >>= y) → (x >>= (y as u8))
- Applies to both normal Solidity expressions and Yul assembly output

SafeCast bitmask:
- Non-standard uint widths (uint24, uint40, uint48, ..., uint248) use
  bitmask instead of no-op cast: uint248(x) → (x & MASK_248_BITS)
- Preserves truncation semantics that Move's non-truncating casts lose

Constant expressions:
- keccak256("string") in constant context → compute hash at transpile
  time using Node.js crypto (sha3-256 approximation for constant IDs)

DLMM results: 22/22 transpile, 0 unsupported in all libraries

+++ tests/output/dlmm/bit_math/sources/bit_math.move
@@ -25,14 +25,14 @@ module 0x1::bit_math {
     public(package) fun closest_bit_right(x: u256, bit: u8): u256 {
         let id = 0u256;
         let shift: u256 = (255u256 - bit);
-        x <<= shift;
+        x <<= (shift as u8);
         if (((x == 0u256))) 115792089237316195423570985008687907853269984665640564039457584007913129639935u256 else (most_significant_bit(x) - shift)
         id
     }
 
     public(package) fun closest_bit_left(x: u256, bit: u8): u256 {
         let id = 0u256;
-        x >>= bit;
+        x >>= (bit as u8);
         if (((x == 0u256))) 115792089237316195423570985008687907853269984665640564039457584007913129639935u256 else (least_significant_bit(x) + bit)
         id
     }
